requirements.txt
dsmr-parser==0.33
influxdb-client==1.31.0
pyserial==3.5
